+++ CHANGELOG.md
@@ -9,11 +9,14 @@
 
 ## Tagged Releases
 
-- [unreleased](https://github.com/Seika139/repo-sync/compare/v0.1.1...HEAD)
+- [unreleased](https://github.com/Seika139/repo-sync/compare/v0.2.0...HEAD)
+- [0.2.0](https://github.com/Seika139/repo-sync/compare/v0.1.1...v0.2.0)
 - [0.1.1](https://github.com/Seika139/repo-sync/compare/v0.1.0...v0.1.1)
 - [0.1.0](https://github.com/Seika139/repo-sync/releases/tag/v0.1.0)
 
 ## [Unreleased]
+
+## [0.2.0] - 2026-05-12
 
 ### Added